refactor: consolidate coverage-chasing tests into behavioral tests (closes #18)
Delete 6 test files that existed primarily to hit coverage branches
rather than assert meaningful parser/server behavior:
- test_100_coverage.py
- test_100_percent_coverage.py
- test_final_100_coverage.py
- test_final_missing_coverage.py
- test_full_coverage.py
- test_missing_coverage.py

Genuinely unique test behaviors from those files were consolidated into:
- tests/test_parser_behavior.py: unknown keyword handling, case-insensitive
  lookup, geometry edge cases, percent block edge cases, empty input
- tests/test_server_behavior.py: feature registration, hover/completion/
  diagnostic/code-action behaviors, document events, geometry detection

Also fix test_full_coverage_enhanced.py: 6 tests that depended on
PropertyMock side effects from deleted tests now use proper workspace
mocking, making them pass independently of test execution order.

Test count: 353 -> 265 (removed 88 redundant tests, preserved all
unique behavioral coverage).
