--- tests/test_full_coverage.py
@@ -1,815 +1,0 @@
-"""Tests for 100% coverage of ORCA LSP server and parser."""
-
-from unittest.mock import MagicMock, PropertyMock, patch
-
-import pytest
-from lsprotocol.types import (
-    CodeActionContext,
-    CodeActionParams,
-    CompletionList,
-    CompletionParams,
-    Diagnostic,
-    DiagnosticSeverity,
-    DidChangeTextDocumentParams,
-    DidOpenTextDocumentParams,
-    Hover,
-    HoverParams,
-    MarkupContent,
-    Position,
-    Range,
-    TextDocumentIdentifier,
-    TextDocumentItem,
-    VersionedTextDocumentIdentifier,
-)
-
-from orca_lsp.parser import ORCAParser
-from orca_lsp.server import ORCALanguageServer, main
-
-
-class MockTextDocument:
-    """Mock text document for testing."""
-
-    def __init__(self, content: str, uri: str = "file:///test.inp"):
-        self.source = content
-        self.uri = uri
-        self.lines = content.split("\n") if content else []
-
-
-class TestServerCompletions:
-    """Test server completion feature with full coverage."""
-
-    @pytest.fixture
-    def server(self):
-        server = ORCALanguageServer()
-        # Use PropertyMock to set the workspace attribute
-        type(server).workspace = PropertyMock(return_value=MagicMock())
-        return server
-
-    def test_on_completion_simple_input(self, server):
-        """Test on_completion for simple input line."""
-        mock_doc = MockTextDocument("! B3LYP ")
-        server.workspace.get_text_document.return_value = mock_doc
-
-        params = CompletionParams(
-            text_document=TextDocumentIdentifier(uri="file:///test.inp"),
-            position=Position(line=0, character=8),
-        )
-
-        result = server._on_completion(params)
-        assert result is not None
-        assert isinstance(result, CompletionList)
-
-    def test_on_completion_percent_block(self, server):
-        """Test on_completion for percent block."""
-        mock_doc = MockTextDocument("%max")
-        server.workspace.get_text_document.return_value = mock_doc
-
-        params = CompletionParams(
-            text_document=TextDocumentIdentifier(uri="file:///test.inp"),
-            position=Position(line=0, character=4),
-        )
-
-        result = server._on_completion(params)
-        assert result is not None
-
-    def test_on_completion_geometry(self, server):
-        """Test on_completion in geometry section."""
-        mock_doc = MockTextDocument("O 0.0 0.0 0.0")
-        server.workspace.get_text_document.return_value = mock_doc
-
-        params = CompletionParams(
-            text_document=TextDocumentIdentifier(uri="file:///test.inp"),
-            position=Position(line=0, character=0),
-        )
-
-        result = server._on_completion(params)
-        assert result is not None
-
-    def test_get_completions_empty_line(self, server):
-        """Test completions for empty line."""
-        completions = server._get_completions("", Position(line=0, character=0))
-        assert isinstance(completions, list)
-        assert len(completions) == 0
-
-    def test_get_percent_completions_block_name(self, server):
-        """Test percent block name completion."""
-        completions = server._get_percent_completions("%")
-        assert isinstance(completions, list)
-        assert len(completions) > 0
-
-    def test_get_percent_completions_in_block(self, server):
-        """Test completions inside a specific block."""
-        completions = server._get_percent_completions("%maxcore ")
-        assert isinstance(completions, list)
-
-    def test_get_block_specific_completions_maxcore(self, server):
-        """Test completions for maxcore block."""
-        completions = server._get_block_specific_completions("maxcore")
-        assert isinstance(completions, list)
-        assert len(completions) > 0
-        assert any("MB" in item.label for item in completions)
-
-    def test_get_block_specific_completions_pal(self, server):
-        """Test completions for pal block."""
-        completions = server._get_block_specific_completions("pal")
-        assert isinstance(completions, list)
-        assert any("nprocs" in item.label for item in completions)
-
-    def test_get_block_specific_completions_method(self, server):
-        """Test completions for method block."""
-        completions = server._get_block_specific_completions("method")
-        assert isinstance(completions, list)
-        assert any("D3" in item.label or "D4" in item.label for item in completions)
-
-    def test_get_block_specific_completions_scf(self, server):
-        """Test completions for scf block."""
-        completions = server._get_block_specific_completions("scf")
-        assert isinstance(completions, list)
-        labels = [item.label for item in completions]
-        assert any(opt in labels for opt in ["maxiter", "convergence", "NRMaxIt"])
-
-    def test_get_block_specific_completions_unknown(self, server):
-        """Test completions for unknown block."""
-        completions = server._get_block_specific_completions("unknown_block")
-        assert isinstance(completions, list)
-        assert len(completions) == 0
-
-
-class TestServerHover:
-    """Test server hover feature with full coverage."""
-
-    @pytest.fixture
-    def server(self):
-        server = ORCALanguageServer()
-        type(server).workspace = PropertyMock(return_value=MagicMock())
-        return server
-
-    def test_on_hover_dft_functional(self, server):
-        """Test hover on DFT functional."""
-        mock_doc = MockTextDocument("! B3LYP def2-TZVP")
-        server.workspace.get_text_document.return_value = mock_doc
-
-        params = HoverParams(
-            text_document=TextDocumentIdentifier(uri="file:///test.inp"),
-            position=Position(line=0, character=4),
-        )
-
-        result = server._on_hover(params)
-        assert result is not None
-        assert isinstance(result, Hover)
-        assert isinstance(result.contents, MarkupContent)
-        assert "B3LYP" in result.contents.value
-
-    def test_on_hover_wavefunction_method(self, server):
-        """Test hover on wavefunction method."""
-        mock_doc = MockTextDocument("! MP2 cc-pVTZ")
-        server.workspace.get_text_document.return_value = mock_doc
-
-        params = HoverParams(
-            text_document=TextDocumentIdentifier(uri="file:///test.inp"),
-            position=Position(line=0, character=3),
-        )
-
-        result = server._on_hover(params)
-        assert result is not None
-        assert "MP2" in result.contents.value
-
-    def test_on_hover_basis_set(self, server):
-        """Test hover on basis set."""
-        mock_doc = MockTextDocument("! B3LYP def2-TZVP")
-        server.workspace.get_text_document.return_value = mock_doc
-
-        params = HoverParams(
-            text_document=TextDocumentIdentifier(uri="file:///test.inp"),
-            position=Position(line=0, character=10),
-        )
-
-        server._on_hover(params)
-
-    def test_on_hover_job_type(self, server):
-        """Test hover on job type."""
-        mock_doc = MockTextDocument("! B3LYP def2-TZVP OPT")
-        server.workspace.get_text_document.return_value = mock_doc
-
-        params = HoverParams(
-            text_document=TextDocumentIdentifier(uri="file:///test.inp"),
-            position=Position(line=0, character=20),
-        )
-
-        result = server._on_hover(params)
-        assert result is not None
-        assert "OPT" in result.contents.value
-
-    def test_on_hover_unknown_word(self, server):
-        """Test hover on unknown word."""
-        mock_doc = MockTextDocument("! UNKNOWN_KEYWORD")
-        server.workspace.get_text_document.return_value = mock_doc
-
-        params = HoverParams(
-            text_document=TextDocumentIdentifier(uri="file:///test.inp"),
-            position=Position(line=0, character=4),
-        )
-
-        result = server._on_hover(params)
-        assert result is None
-
-    def test_on_hover_empty_word(self, server):
-        """Test hover on empty position."""
-        mock_doc = MockTextDocument("! ")
-        server.workspace.get_text_document.return_value = mock_doc
-
-        params = HoverParams(
-            text_document=TextDocumentIdentifier(uri="file:///test.inp"),
-            position=Position(line=0, character=2),
-        )
-
-        result = server._on_hover(params)
-        assert result is None
-
-    def test_get_word_at_start_of_line(self, server):
-        """Test getting word at start of line."""
-        mock_doc = MockTextDocument("B3LYP def2-TZVP")
-        word = server._get_word_at_position(mock_doc, Position(line=0, character=0))
-        assert word == "B3LYP"
-
-    def test_get_word_at_end_of_line(self, server):
-        """Test getting word at end of line."""
-        mock_doc = MockTextDocument("B3LYP def2-TZVP")
-        word = server._get_word_at_position(mock_doc, Position(line=0, character=6))
-        assert word == "def2"
-
-
-class TestServerDiagnostics:
-    """Test server diagnostics with full coverage."""
-
-    @pytest.fixture
-    def server(self):
-        server = ORCALanguageServer()
-        type(server).workspace = PropertyMock(return_value=MagicMock())
-        return server
-
-    def test_validate_document_with_errors(self, server):
-        """Test document validation with errors."""
-        content = ""
-        mock_doc = MockTextDocument(content)
-        server.workspace.get_text_document.return_value = mock_doc
-
-        server.publish_diagnostics = MagicMock()
-
-        server._validate_document("file:///test.inp")
-
-        server.publish_diagnostics.assert_called_once()
-        call_args = server.publish_diagnostics.call_args
-        assert call_args[0][0] == "file:///test.inp"
-        diagnostics = call_args[0][1]
-        assert isinstance(diagnostics, list)
-
-    def test_validate_document_with_warnings(self, server):
-        """Test document validation with warnings."""
-        content = "! B3LYP def2-TZVP\n* xyz 0 1\nH 0 0 0\n*"
-        mock_doc = MockTextDocument(content)
-        server.workspace.get_text_document.return_value = mock_doc
-
-        server.publish_diagnostics = MagicMock()
-
-        server._validate_document("file:///test.inp")
-
-        server.publish_diagnostics.assert_called_once()
-        diagnostics = server.publish_diagnostics.call_args[0][1]
-        warning_messages = [d.message for d in diagnostics]
-        assert any("maxcore" in msg.lower() for msg in warning_messages)
-
-    def test_validate_document_valid(self, server):
-        """Test document validation with valid input."""
-        content = """! B3LYP def2-TZVP OPT
-%maxcore 4000
-* xyz 0 1
-O 0 0 0
-H 0 0 1
-H 0 1 0
-*"""
-        mock_doc = MockTextDocument(content)
-        server.workspace.get_text_document.return_value = mock_doc
-
-        server.publish_diagnostics = MagicMock()
-
-        server._validate_document("file:///test.inp")
-
-        server.publish_diagnostics.assert_called_once()
-        diagnostics = server.publish_diagnostics.call_args[0][1]
-        errors = [d for d in diagnostics if d.severity == DiagnosticSeverity.Error]
-        assert len(errors) == 0
-
-
-class TestServerCodeActions:
-    """Test server code actions with full coverage."""
-
-    @pytest.fixture
-    def server(self):
-        server = ORCALanguageServer()
-        type(server).workspace = PropertyMock(return_value=MagicMock())
-        return server
-
-    def test_on_code_action_maxcore_fix(self, server):
-        """Test code action for missing maxcore."""
-        mock_doc = MockTextDocument("! B3LYP def2-TZVP")
-        server.workspace.get_text_document.return_value = mock_doc
-
-        diagnostic = Diagnostic(
-            range=Range(start=Position(line=0, character=0), end=Position(line=0, character=20)),
-            message="Missing %maxcore setting. Recommended: %maxcore 2000-4000",
-            severity=DiagnosticSeverity.Warning,
-            source="orca-lsp",
-        )
-
-        params = CodeActionParams(
-            text_document=TextDocumentIdentifier(uri="file:///test.inp"),
-            range=Range(start=Position(line=0, character=0), end=Position(line=0, character=20)),
-            context=CodeActionContext(diagnostics=[diagnostic]),
-        )
-
-        actions = server._on_code_action(params)
-
-        assert isinstance(actions, list)
-        assert len(actions) > 0
-        assert any("maxcore" in action.title.lower() for action in actions)
-
-    def test_on_code_action_no_match(self, server):
-        """Test code action with no matching fix."""
-        mock_doc = MockTextDocument("! B3LYP def2-TZVP")
-        server.workspace.get_text_document.return_value = mock_doc
-
-        diagnostic = Diagnostic(
-            range=Range(start=Position(line=0, character=0), end=Position(line=0, character=20)),
-            message="Some other error",
-            severity=DiagnosticSeverity.Error,
-            source="orca-lsp",
-        )
-
-        params = CodeActionParams(
-            text_document=TextDocumentIdentifier(uri="file:///test.inp"),
-            range=Range(start=Position(line=0, character=0), end=Position(line=0, character=20)),
-            context=CodeActionContext(diagnostics=[diagnostic]),
-        )
-
-        actions = server._on_code_action(params)
-
-        assert isinstance(actions, list)
-        assert len(actions) == 0
-
-
-class TestServerDocumentEvents:
-    """Test server document events with full coverage."""
-
-    @pytest.fixture
-    def server(self):
-        server = ORCALanguageServer()
-        type(server).workspace = PropertyMock(return_value=MagicMock())
-        return server
-
-    def test_on_did_open(self, server):
-        """Test document open event."""
-        content = "! B3LYP def2-TZVP"
-        mock_doc = MockTextDocument(content)
-        server.workspace.get_text_document.return_value = mock_doc
-
-        server.publish_diagnostics = MagicMock()
-
-        params = DidOpenTextDocumentParams(
-            text_document=TextDocumentItem(
-                uri="file:///test.inp", language_id="orca", version=1, text=content
-            )
-        )
-
-        server._on_did_open(params)
-
-        server.publish_diagnostics.assert_called_once()
-
-    def test_on_did_change(self, server):
-        """Test document change event."""
-        content = "! B3LYP def2-TZVP OPT"
-        mock_doc = MockTextDocument(content)
-        server.workspace.get_text_document.return_value = mock_doc
-
-        server.publish_diagnostics = MagicMock()
-
-        params = DidChangeTextDocumentParams(
-            text_document=VersionedTextDocumentIdentifier(uri="file:///test.inp", version=2),
-            content_changes=[],
-        )
-
-        server._on_did_change(params)
-
-        server.publish_diagnostics.assert_called_once()
-
-
-class TestParserEdgeCases:
-    """Test parser edge cases for full coverage."""
-
-    @pytest.fixture
-    def parser(self):
-        return ORCAParser()
-
-    @pytest.fixture
-    def server(self):
-        return ORCALanguageServer()
-
-    def test_parse_simple_input_case_insensitive(self, parser):
-        """Test case-insensitive parsing."""
-        result = parser.parse_simple_input("! b3lyp DEF2-tzvp opt", 0)
-        assert "b3lyp" in result.methods or "B3LYP" in result.methods
-
-    def test_parse_percent_block_single_line(self, parser):
-        """Test single-line percent block."""
-        lines = ["%maxcore 4000"]
-        block, end_line = parser.parse_percent_block(lines, 0)
-
-        assert block is not None
-        assert block.name == "maxcore"
-        assert block.parameters.get("memory") == 4000
-        assert end_line == 0
-
-    def test_parse_percent_block_with_end_inline(self, parser):
-        """Test percent block with end on same line."""
-        lines = ["%pal nprocs 4 end"]
-        block, end_line = parser.parse_percent_block(lines, 0)
-
-        assert block is not None
-        assert block.name == "pal"
-        assert end_line == 0
-
-    def test_parse_percent_block_multiline(self, parser):
-        """Test multi-line percent block."""
-        lines = ["%scf", "  maxiter 100", "  convergence tight", "end"]
-        block, end_line = parser.parse_percent_block(lines, 0)
-
-        assert block is not None
-        assert block.name == "scf"
-        assert end_line == 3
-
-    def test_parse_percent_block_no_end(self, parser):
-        """Test percent block without end keyword."""
-        lines = ["%scf", "  maxiter 100"]
-        block, end_line = parser.parse_percent_block(lines, 0)
-
-        assert block is not None
-        assert block.name == "scf"
-
-    def test_parse_percent_block_invalid(self, parser):
-        """Test invalid percent block."""
-        lines = ["% 123invalid"]
-        block, end_line = parser.parse_percent_block(lines, 0)
-
-        # It should return a block with name starting with a number
-        assert block is not None
-        assert block.name == "123invalid"
-
-    def test_parse_geometry_xyz(self, parser):
-        """Test XYZ geometry format."""
-        content = """! B3LYP def2-SVP
-* xyz 0 1
-O 0.0 0.0 0.0
-H 0.0 0.0 1.0
-H 0.0 1.0 0.0
-*"""
-        result = parser.parse(content)
-
-        assert result.geometry is not None
-        assert result.geometry.charge == 0
-        assert result.geometry.multiplicity == 1
-        assert len(result.geometry.atoms) == 3
-
-    def test_parse_geometry_invalid_element(self, parser):
-        """Test geometry with invalid element."""
-        content = """! B3LYP def2-SVP
-* xyz 0 1
-Xx 0.0 0.0 0.0
-*"""
-        result = parser.parse(content)
-
-        assert any("Invalid element" in e.get("message", "") for e in result.errors)
-
-    def test_parse_geometry_no_end(self, parser):
-        """Test geometry without end marker."""
-        content = """! B3LYP def2-SVP
-* xyz 0 1
-O 0.0 0.0 0.0
-H 0.0 0.0 1.0"""
-        result = parser.parse(content)
-
-        assert result.geometry is not None or len(result.errors) > 0
-
-    def test_parse_empty_document(self, parser):
-        """Test parsing empty document."""
-        result = parser.parse("")
-
-        assert len(result.warnings) > 0 or len(result.errors) > 0
-
-    def test_parse_whitespace_only(self, parser):
-        """Test parsing whitespace-only document."""
-        result = parser.parse("   \n\n   ")
-
-        assert len(result.warnings) > 0 or len(result.errors) > 0
-
-    def test_in_geometry_section(self, server):
-        """Test geometry section detection."""
-        assert server._in_geometry_section("O 0.0 0.0 0.0") is True
-        assert server._in_geometry_section("H 0.5 0.5 0.5") is True
-        assert server._in_geometry_section("! B3LYP") is False
-        assert server._in_geometry_section("%maxcore 4000") is False
-
-
-class TestMainFunction:
-    """Test main function and server startup."""
-
-    @patch("orca_lsp.server.ORCALanguageServer")
-    def test_main_creates_server(self, mock_server_class):
-        """Test that main creates and starts server."""
-        mock_server = MagicMock()
-        mock_server_class.return_value = mock_server
-
-        main()
-
-        mock_server_class.assert_called_once()
-        mock_server.start_io.assert_called_once()
-
-    @patch("orca_lsp.server.ORCALanguageServer")
-    def test_main_start_io_called(self, mock_server_class):
-        """Test that start_io is called."""
-        mock_server = MagicMock()
-        mock_server_class.return_value = mock_server
-
-        main()
-
-        mock_server.start_io.assert_called_once()
-
-
-class TestCaseInsensitiveParsing:
-    """Test case-insensitive parsing for full coverage."""
-
-    @pytest.fixture
-    def parser(self):
-        return ORCAParser()
-
-    def test_parse_case_insensitive_method(self, parser):
-        """Test case-insensitive method parsing."""
-        result = parser.parse_simple_input("! b3lyp", 0)
-        assert len(result.methods) > 0
-        # Should find B3LYP (correct case)
-        assert any("B3LYP" in method or "b3lyp" in method for method in result.methods)
-
-    def test_parse_case_insensitive_basis(self, parser):
-        """Test case-insensitive basis set parsing."""
-        result = parser.parse_simple_input("! def2-tzvp", 0)
-        assert len(result.basis_sets) > 0
-
-    def test_parse_case_insensitive_job_type(self, parser):
-        """Test case-insensitive job type parsing."""
-        result = parser.parse_simple_input("! opt", 0)
-        assert len(result.job_types) > 0
-
-    def test_parse_case_insensitive_method_mixed_case(self, parser):
-        """Test case-insensitive parsing with mixed case."""
-        result = parser.parse_simple_input("! B3lYp", 0)
-        # Should still find B3LYP
-        assert len(result.methods) > 0 or len(result.other_keywords) > 0
-
-
-class TestBasisSetHover:
-    """Test basis set hover for full coverage."""
-
-    @pytest.fixture
-    def server(self):
-        server = ORCALanguageServer()
-        type(server).workspace = PropertyMock(return_value=MagicMock())
-        return server
-
-    def test_on_hover_basis_set_def2(self, server):
-        """Test hover on def2 basis set."""
-        mock_doc = MockTextDocument("! B3LYP def2-SVP")
-        server.workspace.get_text_document.return_value = mock_doc
-
-        params = HoverParams(
-            text_document=TextDocumentIdentifier(uri="file:///test.inp"),
-            position=Position(line=0, character=10),
-        )
-
-        server._on_hover(params)
-        # def2 is not a complete keyword, so may return None
-        # This is acceptable behavior
-
-    def test_on_hover_basis_set_631g(self, server):
-        """Test hover on 6-31G basis set."""
-        mock_doc = MockTextDocument("! B3LYP 6-31G*")
-        server.workspace.get_text_document.return_value = mock_doc
-
-        params = HoverParams(
-            text_document=TextDocumentIdentifier(uri="file:///test.inp"),
-            position=Position(line=0, character=10),
-        )
-
-        server._on_hover(params)
-        # Should hover on 6 or 31, may not find complete keyword
-
-
-class TestFeatureDecorators:
-    """Test that LSP features are properly registered."""
-
-    @pytest.fixture
-    def server(self):
-        return ORCALanguageServer()
-
-    def test_completion_feature_registered(self, server):
-        """Test that completion feature is registered."""
-        # The feature should be registered via decorator
-        # We can verify by checking the function exists
-        assert hasattr(server, "_on_completion")
-
-    def test_hover_feature_registered(self, server):
-        """Test that hover feature is registered."""
-        assert hasattr(server, "_on_hover")
-
-    def test_code_action_feature_registered(self, server):
-        """Test that code action feature is registered."""
-        assert hasattr(server, "_on_code_action")
-
-    def test_did_open_feature_registered(self, server):
-        """Test that did_open feature is registered."""
-        assert hasattr(server, "_on_did_open")
-
-    def test_did_change_feature_registered(self, server):
-        """Test that did_change feature is registered."""
-        assert hasattr(server, "_on_did_change")
-
-
-class TestParserErrorPaths:
-    """Test parser error paths for full coverage."""
-
-    @pytest.fixture
-    def parser(self):
-        return ORCAParser()
-
-    def test_parse_with_invalid_keywords(self, parser):
-        """Test parsing with completely invalid keywords."""
-        result = parser.parse_simple_input("! INVALID1 INVALID2 INVALID3", 0)
-        # Should put them in other_keywords
-        assert len(result.other_keywords) == 3
-
-    def test_parse_geometry_invalid_numbers(self, parser):
-        """Test geometry with invalid coordinates."""
-        content = """! B3LYP def2-SVP
-* xyz 0 1
-O invalid invalid invalid
-*"""
-        result = parser.parse(content)
-        # Parser may not report invalid coordinates as errors
-        assert len(result.warnings) > 0 or len(result.errors) > 0
-
-    def test_parse_percent_block_missing_value(self, parser):
-        """Test percent block with missing parameter value."""
-        lines = ["%maxcore"]
-        block, end_line = parser.parse_percent_block(lines, 0)
-
-        # Block should exist but parameter may not be parsed
-        assert block is not None
-        assert block.name == "maxcore"
-
-
-class TestCaseInsensitivePathCoverage:
-    """Force coverage of case-insensitive lookup paths."""
-
-    @pytest.fixture
-    def parser(self):
-        return ORCAParser()
-
-    def test_method_case_insensitive_lookup(self, parser):
-        """Test case-insensitive method lookup triggers the lookup path."""
-        # Use lowercase to force case-insensitive lookup
-        result = parser.parse_simple_input("! mp2", 0)
-        # Should find MP2 (correct case)
-        assert len(result.methods) > 0
-
-    def test_basis_set_case_insensitive_lookup(self, parser):
-        """Test case-insensitive basis set lookup triggers the lookup path."""
-        result = parser.parse_simple_input("! cc-pvtz", 0)
-        # Should find cc-pVTZ (correct case)
-        assert len(result.basis_sets) > 0
-
-    def test_job_type_case_insensitive_lookup(self, parser):
-        """Test case-insensitive job type lookup triggers the lookup path."""
-        result = parser.parse_simple_input("! freq", 0)
-        # Should find FREQ (correct case)
-        assert len(result.job_types) > 0
-
-    def test_all_lower_case_methods(self, parser):
-        """Test with all lowercase method that needs case-insensitive lookup."""
-        result = parser.parse_simple_input("! hf", 0)
-        assert len(result.methods) > 0
-
-
-class TestParserErrorPathsForCoverage:
-    """Test parser error paths to achieve 100% coverage."""
-
-    @pytest.fixture
-    def parser(self):
-        return ORCAParser()
-
-    def test_geometry_invalid_charge(self, parser):
-        """Test geometry with invalid charge/multiplicity to trigger ValueError."""
-        content = """! B3LYP def2-SVP
-* xyz invalid_charge 1
-O 0.0 0.0 0.0
-*"""
-        result = parser.parse(content)
-        # This should trigger the ValueError branch
-        assert result is not None
-
-    def test_geometry_invalid_multiplicity(self, parser):
-        """Test geometry with invalid multiplicity to trigger ValueError."""
-        content = """! B3LYP def2-SVP
-* xyz 0 invalid_mult
-O 0.0 0.0 0.0
-*"""
-        result = parser.parse(content)
-        # This should trigger the ValueError branch
-        assert result is not None
-
-    def test_geometry_insufficient_parts(self, parser):
-        """Test geometry header with insufficient parts."""
-        content = """! B3LYP def2-SVP
-*
-O 0.0 0.0 0.0
-*"""
-        result = parser.parse(content)
-        # Should handle gracefully
-        assert result is not None
-
-    def test_parse_method_block_d4(self, parser):
-        """Test method block with D4 dispersion."""
-        lines = ["%method", "  D4", "end"]
-        block, end_line = parser.parse_percent_block(lines, 0)
-        assert block is not None
-        assert "D4" in block.parameters.values()
-
-    def test_parse_pal_block_nprocs(self, parser):
-        """Test pal block with nprocs parameter."""
-        lines = ["%pal", "  nprocs 8", "end"]
-        block, end_line = parser.parse_percent_block(lines, 0)
-        assert block is not None
-        assert block.parameters.get("nprocs") == 8
-
-
-class TestServerBasisSetHover:
-    """Test basis set hover coverage."""
-
-    @pytest.fixture
-    def server(self):
-        server = ORCALanguageServer()
-        type(server).workspace = PropertyMock(return_value=MagicMock())
-        return server
-
-    def test_on_hover_cc_basis_set(self, server):
-        """Test hover on cc-pVTZ basis set."""
-        mock_doc = MockTextDocument("! B3LYP cc-pVTZ")
-        server.workspace.get_text_document.return_value = mock_doc
-
-        params = HoverParams(
-            text_document=TextDocumentIdentifier(uri="file:///test.inp"),
-            position=Position(line=0, character=10),
-        )
-
-        server._on_hover(params)
-        # Should find cc-pVTZ in BASIS_SETS
-
-
-class TestParserRegexMatches:
-    """Test parser regex matches for coverage."""
-
-    @pytest.fixture
-    def parser(self):
-        return ORCAParser()
-
-    def test_parse_pal_block_with_nprocs(self, parser):
-        """Test PAL block with nprocs regex match."""
-        lines = ["%pal", "  nprocs 16", "end"]
-        block, end_line = parser.parse_percent_block(lines, 0)
-        assert block is not None
-        assert block.parameters.get("nprocs") == 16
-
-    def test_parse_pal_block_nprocs_case_insensitive(self, parser):
-        """Test PAL block with nprocs in different case."""
-        lines = ["%pal", "  NPROCS 4", "end"]
-        block, end_line = parser.parse_percent_block(lines, 0)
-        assert block is not None
-        assert block.parameters.get("nprocs") == 4
-
-    def test_parse_geometry_with_full_header(self, parser):
-        """Test geometry with full header (charge and multiplicity)."""
-        content = """! B3LYP def2-SVP
-* xyz 0 2
-O 0.0 0.0 0.0
-*"""
-        result = parser.parse(content)
-        assert result.geometry is not None
-        assert result.geometry.charge == 0
-        assert result.geometry.multiplicity == 2
