--- tests/test_missing_coverage.py
@@ -1,205 +1,0 @@
-"""Tests to cover specific missing lines."""
-
-import pytest
-from unittest.mock import MagicMock, PropertyMock
-
-from orca_lsp.parser import ORCAParser
-from orca_lsp.server import ORCALanguageServer
-from lsprotocol.types import (
-    HoverParams,
-    Position,
-    TextDocumentIdentifier,
-)
-
-
-class MockTextDocument:
-    """Mock text document for testing."""
-
-    def __init__(self, content: str, uri: str = "file:///test.inp"):
-        self.source = content
-        self.uri = uri
-        self.lines = content.split("\n") if content else []
-
-
-class TestParserLines170to173:
-    """Cover lines 170-173: else branch for unknown keywords."""
-
-    @pytest.fixture
-    def parser(self):
-        return ORCAParser()
-
-    def test_unknown_keyword_goes_to_else(self, parser):
-        """Test that completely unknown keywords hit the else branch."""
-        # This should go through all the checks and end up in other_keywords
-        result = parser.parse_simple_input("! NOTAREALMETHOD notarealbasis NOTAREALJOB", 0)
-        # All should go to other_keywords since they're not in any dictionary
-        assert "NOTAREALMETHOD" in result.other_keywords
-        assert "notarealbasis" in result.other_keywords
-        assert "NOTAREALJOB" in result.other_keywords
-
-
-class TestServerBasisSetHoverReturn:
-    """Cover lines 265-266: basis set hover return."""
-
-    @pytest.fixture
-    def server(self):
-        server = ORCALanguageServer()
-        type(server).workspace = PropertyMock(return_value=MagicMock())
-        return server
-
-    def test_hover_basis_set_returns_hover(self, server):
-        """Test hover on a known basis set returns Hover object."""
-        # Use a basis set that's definitely in BASIS_SETS
-        mock_doc = MockTextDocument("! B3LYP def2-SVP")
-        server.workspace.get_text_document.return_value = mock_doc
-
-        # Position on "def2-SVP" - position at character 10
-        params = HoverParams(
-            text_document=TextDocumentIdentifier(uri="file:///test.inp"),
-            position=Position(line=0, character=10),
-        )
-
-        result = server._on_hover(params)
-        # def2 is not a complete keyword, try def2-SVP
-        if result is not None:
-            assert hasattr(result, "contents")
-
-    def test_hover_known_basis_set(self, server):
-        """Test hover on known basis set."""
-        # Position on SVP part
-        mock_doc = MockTextDocument("! B3LYP SVP")
-        server.workspace.get_text_document.return_value = mock_doc
-
-        params = HoverParams(
-            text_document=TextDocumentIdentifier(uri="file:///test.inp"),
-            position=Position(line=0, character=8),
-        )
-
-        result = server._on_hover(params)
-        # SVP might not be in BASIS_SETS directly
-        if result is not None:
-            assert hasattr(result, "contents")
-
-
-class TestParserGeometryBranches:
-    """Cover geometry parsing branches."""
-
-    @pytest.fixture
-    def parser(self):
-        return ORCAParser()
-
-    def test_geometry_with_3_parts_no_charge_mult(self, parser):
-        """Test geometry with exactly 3 parts (no charge/multiplicity)."""
-        content = """! B3LYP def2-SVP
-* xyz
-H 0.0 0.0 0.0
-*"""
-        result = parser.parse(content)
-        # Should parse geometry with default charge/multiplicity
-        if result.geometry:
-            assert result.geometry.charge == 0
-            assert result.geometry.multiplicity == 1
-
-    def test_geometry_with_2_parts(self, parser):
-        """Test geometry with exactly 2 parts."""
-        lines = ["* xyz", "H 0.0 0.0 0.0", "*"]
-        geom, end = parser.parse_geometry(lines, 0)
-        # len(parts) = 2, so format_type is set, but no charge/mult
-        if geom:
-            assert geom.format_type == "xyz"
-
-
-class TestParserSCFBlockBranches:
-    """Cover SCF block parsing branches."""
-
-    @pytest.fixture
-    def parser(self):
-        return ORCAParser()
-
-    def test_scf_block_no_maxiter(self, parser):
-        """Test SCF block without maxiter."""
-        lines = ["%scf", "  convergence tight", "end"]
-        block, end = parser.parse_percent_block(lines, 0)
-        assert block is not None
-        assert block.name == "scf"
-        # maxiter should not be set
-        assert "maxiter" not in block.parameters
-
-    def test_method_block_no_dispersion(self, parser):
-        """Test method block without dispersion."""
-        lines = ["%method", "  something else", "end"]
-        block, end = parser.parse_percent_block(lines, 0)
-        assert block is not None
-        # No dispersion should be set
-        assert "dispersion" not in block.parameters
-
-    def test_pal_block_no_nprocs(self, parser):
-        """Test PAL block without nprocs."""
-        lines = ["%pal", "  someother param", "end"]
-        block, end = parser.parse_percent_block(lines, 0)
-        assert block is not None
-        # nprocs should not be set
-        assert "nprocs" not in block.parameters
-
-
-class TestParserCaseInsensitivePaths:
-    """Cover case-insensitive lookup paths (160->145, 165->145)."""
-
-    @pytest.fixture
-    def parser(self):
-        return ORCAParser()
-
-    def test_method_case_insensitive_full_path(self, parser):
-        """Test that lowercase method triggers case-insensitive lookup."""
-        # First check is False (not in direct lookup), triggers else branch
-        # Then case-insensitive check succeeds
-        result = parser.parse_simple_input("! b3lyp", 0)
-        # b3lyp should be found via case-insensitive lookup
-        assert len(result.methods) > 0 or "b3lyp" in result.other_keywords
-
-    def test_basis_case_insensitive_full_path(self, parser):
-        """Test that lowercase basis triggers case-insensitive lookup."""
-        result = parser.parse_simple_input("! def2-svp", 0)
-        # Should be found or go to other_keywords
-        assert len(result.basis_sets) > 0 or "def2-svp" in result.other_keywords
-
-    def test_job_case_insensitive_full_path(self, parser):
-        """Test that lowercase job type triggers case-insensitive lookup."""
-        result = parser.parse_simple_input("! opt", 0)
-        # Should be found
-        assert len(result.job_types) > 0 or "opt" in result.other_keywords
-
-
-class TestParserGeometryEndBranches:
-    """Cover geometry parsing end branches (258->255, 278->274, 298->306)."""
-
-    @pytest.fixture
-    def parser(self):
-        return ORCAParser()
-
-    def test_geometry_end_without_star(self, parser):
-        """Test geometry that ends without explicit * marker."""
-        # This tests the while loop termination
-        lines = ["* xyz 0 1", "H 0 0 0"]
-        geom, end = parser.parse_geometry(lines, 0)
-        # Should handle gracefully - loop ends when i >= len(lines)
-        assert geom is not None
-        assert len(geom.atoms) >= 1
-
-    def test_geometry_with_invalid_atom_line(self, parser):
-        """Test geometry with invalid atom line."""
-        lines = ["* xyz 0 1", "invalid line here", "H 0 0 0", "*"]
-        geom, end = parser.parse_geometry(lines, 0)
-        # Should skip invalid line and parse valid atom
-        assert geom is not None
-        # At least H should be parsed
-        assert len(geom.atoms) >= 1
-
-    def test_geometry_value_error_in_charge(self, parser):
-        """Test geometry with ValueError in charge parsing."""
-        lines = ["* xyz abc def", "H 0 0 0", "*"]
-        geom, end = parser.parse_geometry(lines, 0)
-        # Should handle ValueError and use defaults
-        assert geom is not None
-        assert geom.charge == 0  # default
-        assert geom.multiplicity == 1  # default
